image-pipeline/pyproject.toml
[project]
name = "framefast-image-pipeline"
version = "0.1.0"
description = "Image processing pipeline for FrameFast - auto-edit, LUT, upscale"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "colour-science>=0.4.3",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
